Refactored driver data and updated unit tests
